user/requirements.txt
flask
flask_cors
Werkzeug
bson
pika
pymongo
mongoengine
grpcio==1.7.3
grpcio-tools==1.7.3